Split test_grippers.py: keep Franka tests, delete Robotiq tests (fixes #87)
mj_manipulator/tests/test_grippers.py has been broken in CI for
months: it did an unconditional `import geodude_assets` at module
level, but geodude_assets is not a declared dependency of
mj_manipulator (it's a downstream consumer's asset package). CI runs
with UV_NO_SOURCES=true, so the import fails during pytest
collection and the entire test job exits with code 2 before any
tests run. This was masked by admin merges past the failing check on
main for 5+ successive PRs.

#86 added a `pytest.importorskip` Band-Aid to unblock the Franka
polish PR. This commit applies the principled fix from #87:

  - Rename `tests/test_grippers.py` → `tests/test_franka_gripper.py`
  - Keep all Franka gripper tests (14 tests across 4 classes)
  - Remove all Robotiq-specific tests, fixtures, and imports
  - Remove the `pytest.importorskip("geodude_assets")` Band-Aid
  - Remove the `_ROBOTIQ_SCENE` path constant
  - Remove `RobotiqGripper` from the imports

The Robotiq tests now live in geodude/tests/test_robotiq_gripper.py,
where `geodude_assets` (which owns the 2F-140 scene XML) is a
declared dependency. See the paired geodude PR for the moved file.

General principle: tests in package X should only depend on X's
declared dependencies. If a test needs a sibling package's assets,
that test belongs in the sibling, not here.

Verified locally:
  - `grep -r geodude_assets tests/ src/` returns only a docstring reference
  - `uv run pytest tests/ -q` → 319 passed (was 335 with Band-Aid;
    the 16 Robotiq tests moved to geodude's test suite)
  - `uv run ruff check .` and `ruff format --check .` both clean

Fixes personalrobotics/mj_manipulator#87.
